E212: Update MCC list
Op. Bulletin No. 1117 (1.II.2017) has an annexed list of Mobile
Country Codes that includes a number of which are not mentioned
in the MCC/MNC combined list in Op. Bulletin No. 1161 (1.XX.2018),
possibly because the MNCs are not reported to the ITU in a timely
fashion or because the assigned number is not actually used (e.g.,
Vatican City). See:
https://www.itu.int/dms_pub/itu-t/opb/sp/T-SP-OB.1117-2017-OAS-PDF-E.pdf