dfilter: Fixup some scanner comments and patterns
Fixup comment to reflect that unparsed is the only token
type for the pattern, there is no "match" or "no match".

Tighten the CIDR patterns and make sure it cannot match a
rogue "..". This is ambiguous with 80..90 and must be treated
extra carefully to support floating point formats without
leading or terminating digits before or after the point.

Reject the default match as a syntax error. This pattern is
just a pitfall to make testing and debugging harder. Matching a
single arbitrary character is never a valid unparsed string,
except by chance.