BSSAP: Simplify code
Remove now unused defines. Add the BSAP and BSSAP-LE pinos to the tree
as the parent protocol (so that the BSSAP default proto preference can
be selected from the Packet Details and Packet List).

Set the BSSAP+ Decode As preference with the newer Decode As API, so
that we don't need a preference callback.