RTP: Strengthen heuristics
Strengthen the heuristics according to Appendix A.1 of RFC 3550:

Reject if the packet isn't long enough to fit the fixed header, including
the CSRCs if present.
Reject if the packet isn't long enough to fit the extension header.
Reject if the packet isn't long enough to fit the padding, if we have
all the packet.
Reject the payload types reserved for RTCP conflict.

Most of these lead to malformed packet error if dissected anyway,
which can still be done via other methods (SDP, Decode As, etc.)