MKA: Fixing misleading capability value-strings (Bugfix)
The original MACsec capability value strings do not reflect the
IEEE 802.1X specification (2010 or 2020).

For example: IEEE 802.1X says for value 2:
"‘Integrity without confidentiality’ and ‘Integrity and
confidentiality’ with a confidentiality offset of 0"

The packet-mka.c value string for 2 says:
"MACsec Integrity with no confidentiality offset"

The updated value string now shows that integrity and
integrity+confidentiality are supported.