ANSI A I/F: Retrieve stats correctly
The row number to lookup in the stat table is the index retrieved
from my_try_val_to_str_idx, not the original message type.
Ticks the counts in the correct rows of the stats table, and
prevents failed assertions and program halt in stat_tap_ui.c when
getting a message type with a number greater than the number of
rows in the table.


(cherry picked from commit 8b8171d95217fb629cc23fd8004994ee7aa37b18)