GUI: Lower log level of "up and ready" message
This lowers the level of this message from "message" to
"info". This has two side-effects:
  - It is not displayed by default
  - It is printed to stdout instead of stderr.

Some users were depending on this message. Restore this to
the level it had before 05ed76d4. Even though this output is
not considered a stable interface restoring the old behavior
helps them and has no meaningful usability downsides. The
changes in 05ed76d4 were experimental anyway.

Related to #17763.