kafka: have dissect_kafka_string_new() return the display string.
Instead of having it return the information needed to fetch the string
value, just have it return the string to use to display that string, as
that's all its only caller needs.

(Note that the display string has had control characters, etc. escaped,
which is what you want for text that appears in a string displayed in
the protocol details.)