dfilter: Require separators with ISO 8601 time
Require date/time separators when entering a time value, e,g:
    2014-07-04 12:34:56.789+00:00

Separators in the timezone offset are an exception, they are
never mandatory.

This excludes ISO basic format to avoid inputs that could
be entirely numbers indistinguishable from Epoch time, in case
we want to support that in the future.