Signal-Pdu/AUTOSAR I-PduM: CAN-IDs registered as std/ext (bugfix)
Instead of just assuming CAN-IDs > 0x7ff are extended, the new code
checks the EFF_FLAG of the CAN-ID of the Signal_PDU_Binding_CAN and
AUTOSAR_IPDUM_Binding_CAN. This affects registering CAN-IDs with the CAN
dissector as well as config lookups.

This patch changes the config format of Signal_PDU_Binding_CAN and
AUTOSAR_IPDUM_Binding_CAN. CAN-IDs need to include the EFF-Flag now!