import_text+Qt: Don't add impossible dummy headers to wrong encaps
Move the list of encapsulation buttons to a QButtonGroup, which makes
it easy to tell which one of them is activated.

Use that to fix an issue where dummy headers impermissible for an
encapsulation type could be added if the radio button was selected
before switching the encapsulation type. (Only for encapsulations
that allow at least one type, i.e. Ethernet or Wireshark Upper PDU.)