http3: add PRIORITY_UPDATE frame dissection
The HTTP/3 priority update frame is an extension frame defined in
https://datatracker.ietf.org/doc/draft-ietf-httpbis-priority/.
Previously, the HTTP/3 dissector only went as far as reporting the
PRIORITY_UPDATE frame types.

With this change, we extend the HTTP/3 dissection to cover the fields
inside the frame. This makes it easier to see the actual target of the
priority and the value of the hint.