SRTCP: Warn about undecoded, not length error
If we are decoding as SRTCP with encrypted payload but srtcp_info
wasn't set up (e.g. because this was done by Decode As or a heuristic),
then since we can't calculate the length, add the expert info about
an undecoded payload rather than the expert info about an incorrect
length.

Related to #17892