RTCP: Don't always Decode As SRTCP
If RTCP is chosen via Decode As, decode as RTCP.
If SRTCP is chosen via Decode As, decode as SRTCP (assuming that
all packets are encrypted, because we can't tell where the E bit is
in that case.)

If possible [S]RTCP is found via the heuristic dissector, assume either
RTCP or (encrypted) SRTCP based a preference. Perhaps later the heuristic
dissector could be improved to make a better decision.