GitLab CI: Fix Commit Check gitrevision for multiple commits
commit^N means the Nth parent of a commit object (after a merge,
for example.) It is not the same as the Nth generation ancestor
(following only the first parents), which is commit~N, or
alternatively commit^^^^^ (N times), or even commit^1^1^1....
(See 'man gitrevisions' for details.)
When the number of commits is N, we want the Nth generation ancestor.

Fixes error messages like:
fatal: ambiguous argument 'HEAD^3': unknown revision or path not
in the working tree.