afp: fix dissection of afp.access and afp.file_bitmap fields.
Fix regressions in AppleShare dissection by correcting the length
of afp.access and afp.file_bitmap fields to be UINT16 as specified
in the AppleShare protocol specification.

Fix reuse of afp.file_bitmask in CatSearchExt as a 16 and as a 32 bit
value by introducing afp.request_bitmask for the 32 bit Request
Bitmap.

Closes #17907.