RTCP: Reject packets with unsupported version
Have RTCP behave similar to the RTP dissector and reject packets
with a version other than 2 (after adding some entries to the tree
as RTP does.) This is necessary because WebRTC and others often send
STUN or DTLS packets on a port after SDP has set up the RTCP dissector
(with a=rtcp:port or similar). Improves dissection of the files
in #13193 when the key log file is set in the TLS prefs.

Perhaps later a preference could be added, as with RTP.