Qt: Go To Packet field - change from inputMask to QIntValidator
inputMask populates the field with a space for each character
in the mask. Mouse people that click in the field may position
the cursor at the far right of the mask and not be able to enter
a value.
https://www.qtcentre.org/threads/7106-QLineEdit-and-input-mask

Remove the inputMask from the field definition and add a validator.