pflog: Various cleanups.
Correctly handle the length field; it should be rounded up to a multiple
of 4 to determine the full length, it shouldn't just have 3 added to it
under the assumption that length % 4 = 1.

The LEN_PFLOG values refer to OpenBSD releases, so name them
LEN_PFLOG_OPENBSD_{version}., not just BSD.  Give them values that don't
include the padding.

Add FreeBSD and Darwin AF_INET6 values, as this can be used to analyze
non-OpenBSD PF logs.

Add additionaal reason, action, and direction values, with #ifs for
different platforms.  To handle other platforms' PF logs, we'd need a
preference (although what we *really* want are separate LINKTYPE_ values
for different OSes, so the preference would not be needed for newer
files).

Use proto_item_add_item_return_ routines for integral-valued fields.

Show the rule number as decimal, as long as it's an FT_INT32.

Update links for OpenBSD CVS repository, add pfvar.h, and add links to
repositories for other OSes with PF.