pflog: more cleanups.
Fix the URL for the FreeBSD pflog code.

Make the byte order for the UID and the PID an enum, with the default
being *host*-endian, as, from a quick look at the PF code in the OSes
that have it, both the IDs are in the byte order of the host writing
the file.  (This means I need to update libpcap and libwiretap to
byte-swap them when reading a byte-swapped capture file or file section,
as we do with some other pseudo-header fields.  That's next on the
list.)

Add some comments about the signedness of the UID and PID fields.

This is the right way to handle #10202.