file: Optimize Find Packet
Split the match functions in twain, one for case-sensitive and
one for case-insensitive, so we can use memchr to search for the
first byte in the case-sensitive version and ws_mempbrk for the
case-insensitive version. They are highly optimized on most systems
and considerably faster on large files.

Also fix a few issues regarding wide strings, such as false positives
and the length to highlight when matching. Fix #12908